+++ lib/packages/crypto/src/encryptMessage.ts
@@ -1,45 +1,29 @@
-// tslint:disable:no-bitwise
-/** @ignore */
 /** @module crypto */
 
-// FIXME: This implementation is not compatible with current BRS encryption
+/**
+ * Original work Copyright (c) 2019 Burst Apps Team
+ */
 
-import * as CryptoJS from 'crypto-js';
-import {ECKCDSA} from './ec-kcdsa';
 import {Converter} from './converter';
+import {encryptData} from './encryptData';
 import {EncryptedMessage} from '../typings/encryptedMessage';
 
 /**
- * Encrypts a message
- * @param message Message to be encrypted
- * @param recipientPublicKey The recipients public key
- * @param senderPrivateKey The senders private (agreement) key
+ * Encrypts a message (UTF-8 compatible)
+ * @param plaintext Message to be encrypted
+ * @param recipientPublicKeyHex The recipients public key hexadecimal format
+ * @param senderPrivateKeyHex The senders private (agreement) key hexadecimal format
  * @return The encrypted Message
  */
-export function encryptMessage(message: string, recipientPublicKey: string, senderPrivateKey: string): EncryptedMessage {
+export function encryptMessage(plaintext: string, recipientPublicKeyHex: string, senderPrivateKeyHex: string): EncryptedMessage {
 
-    const sharedKey =
-        ECKCDSA.sharedkey(
-            Converter.convertHexStringToByteArray(senderPrivateKey),
-            Converter.convertHexStringToByteArray(recipientPublicKey)
-        );
-
-    const SHARED_KEY_SIZE = sharedKey.length;
-    const randomBytes = CryptoJS.lib.WordArray.random(SHARED_KEY_SIZE);
-    const randomNonce = Converter.convertWordArrayToUint8Array(randomBytes);
-    for (let i = 0; i < SHARED_KEY_SIZE; i++) {
-        sharedKey[i] ^= randomNonce[i];
-    }
-    const nonceHex = randomBytes.toString();
-    const aeskey = Converter.convertByteArrayToHexString(sharedKey);
-
-    const iv = CryptoJS.lib.WordArray.random(16);
-    const encryptedTextBase64 = CryptoJS.AES.encrypt(message, aeskey, {iv}).toString();
+    const data = new Uint8Array(Converter.convertStringToByteArray(plaintext));
+    const encryptedData = encryptData(data, recipientPublicKeyHex, senderPrivateKeyHex);
 
     return {
-        // @ts-ignore
-        data: `${iv.toString(CryptoJS.enc.Base64)}:${encryptedTextBase64}`,
-        nonce: nonceHex,
-        isText: true
+        data: Converter.convertByteArrayToHexString(encryptedData.data),
+        nonce: Converter.convertByteArrayToHexString(encryptedData.nonce),
+        isText: true,
     };
 }
+