+++ lib/packages/crypto/src/index.ts
@@ -1,10 +1,12 @@
 /** @module crypto */
 
-export * from './converter';
+// export * from './converter';
 export * from './encryptAES';
 export * from './decryptAES';
-// export * from './decryptMessage';
-// export * from './encryptMessage';
+export * from './decryptData';
+export * from './encryptData';
+export * from './decryptMessage';
+export * from './encryptMessage';
 export * from './generateMasterKeys';
 export * from './getAccountIdFromPublicKey';
 export * from './hashSHA256';
@@ -14,3 +16,4 @@ export * from './generateSignedTransactionBytes';
 export * from './generateSignature';
 export * from '../typings/keys';
 export * from '../typings/encryptedMessage';
+export * from '../typings/encryptedData';