+++ lib/packages/crypto/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## unreleased
+- Encryption/Decryption compatible with BRS
+- removed big.js 
+
 ## 0.1.0-rc.3
 - Encryption/Decryption of message works (not for old messages yet)
 - reduced package size